adiciona cobertura de testes para models

+++ spec/reset_database.rb
@@ -1,15 +1,5 @@
-def reset_database
-  conn = PG.connect(
-    dbname: 'myapp_test',
-    user: 'postgres',
-    password: 'postgres',
-    host: 'db_test',
-    port: 5432
-  )
-
-  if table_exists?(conn, 'patients')
-    conn.exec("TRUNCATE TABLE patients CASCADE;")
-  end
+def reset_database(conn, table_name)
+  conn.exec("TRUNCATE TABLE patients CASCADE;") if table_exists?(conn, 'patients')
 
   conn.close
 end
